+++ shunting-yard-replicator/src/test/java/com/hotels/shunting/yard/replicator/exec/messaging/MessageReaderAdapterTest.java
@@ -41,8 +41,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 
-import com.expedia.apiary.extensions.receiver.common.messaging.MessageEvent;
-import com.expedia.apiary.extensions.receiver.common.messaging.MessageReader;
 import com.expedia.apiary.extensions.receiver.common.event.AddPartitionEvent;
 import com.expedia.apiary.extensions.receiver.common.event.AlterPartitionEvent;
 import com.expedia.apiary.extensions.receiver.common.event.AlterTableEvent;
@@ -52,9 +50,16 @@ import com.expedia.apiary.extensions.receiver.common.event.DropTableEvent;
 import com.expedia.apiary.extensions.receiver.common.event.EventType;
 import com.expedia.apiary.extensions.receiver.common.event.InsertTableEvent;
 import com.expedia.apiary.extensions.receiver.common.event.ListenerEvent;
+import com.expedia.apiary.extensions.receiver.common.messaging.MessageEvent;
+import com.expedia.apiary.extensions.receiver.common.messaging.MessageReader;
 import com.expedia.apiary.extensions.receiver.sqs.messaging.SqsMessageProperty;
 
+import com.hotels.bdp.circustrain.api.conf.ReplicaTable;
 import com.hotels.bdp.circustrain.api.conf.ReplicationMode;
+import com.hotels.bdp.circustrain.api.conf.SourceTable;
+import com.hotels.shunting.yard.replicator.exec.conf.ShuntingYardTableReplicationsMap;
+import com.hotels.shunting.yard.replicator.exec.conf.ct.ShuntingYardTableReplication;
+import com.hotels.shunting.yard.replicator.exec.conf.ct.ShuntingYardTableReplications;
 import com.hotels.shunting.yard.replicator.exec.event.MetaStoreEvent;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -72,6 +77,8 @@ public class MessageReaderAdapterTest {
   private static final String TEST_TABLE_LOCATION = "s3://table_location";
   private static final String SOURCE_METASTORE_URIS = "thrift://remote_host:9883";
   private static final String RECEIPT_HANDLE = "receiptHandle";
+  private static final String REPLICA_DATABASE = "replica_db";
+  private static final String REPLICA_TABLE = "replica_table";
   private static final Map<String, String> PARAMETERS = ImmutableMap.of(METASTOREURIS.varname, SOURCE_METASTORE_URIS);
 
   private MessageReaderAdapter messageReaderAdapter;
@@ -96,21 +103,101 @@ public class MessageReaderAdapterTest {
     FieldSchema partitionColumn2 = new FieldSchema("column_2", "integer", "");
     FieldSchema partitionColumn3 = new FieldSchema("column_3", "string", "");
 
+    ShuntingYardTableReplication tableReplication = new ShuntingYardTableReplication();
+    SourceTable sourceTable = new SourceTable();
+    sourceTable.setDatabaseName(TEST_DB);
+    sourceTable.setTableName(TEST_TABLE);
+
+    ReplicaTable replicaTable = new ReplicaTable();
+    replicaTable.setDatabaseName(REPLICA_DATABASE);
+    replicaTable.setTableName(REPLICA_TABLE);
+
+    tableReplication.setSourceTable(sourceTable);
+    tableReplication.setReplicaTable(replicaTable);
+
+    List<ShuntingYardTableReplication> tableReplications = new ArrayList<>();
+    tableReplications.add(tableReplication);
+
+    ShuntingYardTableReplications tableReplicationsWrapper = new ShuntingYardTableReplications();
+    tableReplicationsWrapper.setTableReplications(tableReplications);
+
     partitionKeys = ImmutableList.of(partitionColumn1, partitionColumn2, partitionColumn3);
     partitionValues = ImmutableList.of(partition);
-    messageReaderAdapter = new MessageReaderAdapter(messageReader, SOURCE_METASTORE_URIS);
+    messageReaderAdapter = new MessageReaderAdapter(messageReader, SOURCE_METASTORE_URIS,
+        new ShuntingYardTableReplicationsMap(tableReplicationsWrapper));
     when(partition.getValues()).thenReturn(PARTITION_VALUES);
   }
 
   @Test
   public void createTableEvent() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryCreateTableEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryCreateTableEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryCreateTableEvent);
     when(apiaryCreateTableEvent.getEventType()).thenReturn(EventType.CREATE_TABLE);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.CREATE_TABLE, TEST_DB, TEST_TABLE)
+        .builder(EventType.CREATE_TABLE, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
+        .environmentContext(EMPTY_MAP)
+        .parameters(PARAMETERS)
+        .replicationMode(ReplicationMode.FULL)
+        .build();
+
+    MetaStoreEvent actual = messageReaderAdapter.read().get();
+
+    assertMetaStoreEvent(expected, actual);
+  }
+
+  @Test
+  public void createTableEventWhenTableReplicationsAreNotConfigured() {
+    messageReaderAdapter = new MessageReaderAdapter(messageReader, SOURCE_METASTORE_URIS,
+        new ShuntingYardTableReplicationsMap(null));
+
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryCreateTableEvent,
+        Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
+    configureMockedEvent(apiaryCreateTableEvent);
+    when(apiaryCreateTableEvent.getEventType()).thenReturn(EventType.CREATE_TABLE);
+
+    MetaStoreEvent expected = MetaStoreEvent
+        .builder(EventType.CREATE_TABLE, TEST_DB, TEST_TABLE, TEST_DB, TEST_TABLE)
+        .environmentContext(EMPTY_MAP)
+        .parameters(PARAMETERS)
+        .replicationMode(ReplicationMode.FULL)
+        .build();
+
+    MetaStoreEvent actual = messageReaderAdapter.read().get();
+
+    assertMetaStoreEvent(expected, actual);
+  }
+
+  @Test
+  public void createTableEventWhenReplicaTableNameNotProvided() {
+    ShuntingYardTableReplication tableReplication = new ShuntingYardTableReplication();
+    SourceTable sourceTable = new SourceTable();
+    sourceTable.setDatabaseName(TEST_DB);
+    sourceTable.setTableName(TEST_TABLE);
+
+    ReplicaTable replicaTable = new ReplicaTable();
+    replicaTable.setDatabaseName(REPLICA_DATABASE);
+
+    tableReplication.setSourceTable(sourceTable);
+    tableReplication.setReplicaTable(replicaTable);
+
+    List<ShuntingYardTableReplication> tableReplications = new ArrayList<>();
+    tableReplications.add(tableReplication);
+
+    ShuntingYardTableReplications tableReplicationsWrapper = new ShuntingYardTableReplications();
+    tableReplicationsWrapper.setTableReplications(tableReplications);
+
+    messageReaderAdapter = new MessageReaderAdapter(messageReader, SOURCE_METASTORE_URIS,
+        new ShuntingYardTableReplicationsMap(tableReplicationsWrapper));
+
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryCreateTableEvent,
+        Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
+    configureMockedEvent(apiaryCreateTableEvent);
+    when(apiaryCreateTableEvent.getEventType()).thenReturn(EventType.CREATE_TABLE);
+
+    MetaStoreEvent expected = MetaStoreEvent
+        .builder(EventType.CREATE_TABLE, TEST_DB, TEST_TABLE, REPLICA_DATABASE, TEST_TABLE)
         .environmentContext(EMPTY_MAP)
         .parameters(PARAMETERS)
         .replicationMode(ReplicationMode.FULL)
@@ -124,7 +211,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void addPartitionEvent() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAddPartitionEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAddPartitionEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAddPartitionEvent);
 
@@ -133,7 +220,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAddPartitionEvent.getEventType()).thenReturn(EventType.ADD_PARTITION);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ADD_PARTITION, TEST_DB, TEST_TABLE)
+        .builder(EventType.ADD_PARTITION, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .partitionColumns(new ArrayList<String>(PARTITION_KEYS_MAP.keySet()))
         .partitionValues(PARTITION_VALUES)
         .environmentContext(EMPTY_MAP)
@@ -149,7 +236,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void alterPartitionEvent() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterPartitionEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterPartitionEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAlterPartitionEvent);
 
@@ -160,7 +247,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAlterPartitionEvent.getEventType()).thenReturn(EventType.ALTER_PARTITION);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ALTER_PARTITION, TEST_DB, TEST_TABLE)
+        .builder(EventType.ALTER_PARTITION, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .partitionColumns(new ArrayList<String>(PARTITION_KEYS_MAP.keySet()))
         .partitionValues(PARTITION_VALUES)
         .replicationMode(ReplicationMode.FULL)
@@ -176,7 +263,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void alterTableEvent() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterTableEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterTableEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAlterTableEvent);
 
@@ -185,7 +272,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAlterTableEvent.getEventType()).thenReturn(EventType.ALTER_TABLE);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ALTER_TABLE, TEST_DB, TEST_TABLE)
+        .builder(EventType.ALTER_TABLE, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .environmentContext(EMPTY_MAP)
         .parameters(PARAMETERS)
         .replicationMode(ReplicationMode.FULL)
@@ -199,7 +286,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void metadataOnlySyncEventForPartition() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterPartitionEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterPartitionEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAlterPartitionEvent);
 
@@ -210,7 +297,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAlterPartitionEvent.getEventType()).thenReturn(EventType.ALTER_PARTITION);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ALTER_PARTITION, TEST_DB, TEST_TABLE)
+        .builder(EventType.ALTER_PARTITION, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .partitionColumns(new ArrayList<String>(PARTITION_KEYS_MAP.keySet()))
         .partitionValues(PARTITION_VALUES)
         .replicationMode(ReplicationMode.METADATA_UPDATE)
@@ -226,7 +313,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void metadataOnlySyncEventForPartitionWithNullLocations() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterPartitionEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterPartitionEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAlterPartitionEvent);
 
@@ -236,7 +323,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAlterPartitionEvent.getEventType()).thenReturn(EventType.ALTER_PARTITION);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ALTER_PARTITION, TEST_DB, TEST_TABLE)
+        .builder(EventType.ALTER_PARTITION, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .partitionColumns(new ArrayList<String>(PARTITION_KEYS_MAP.keySet()))
         .partitionValues(PARTITION_VALUES)
         .replicationMode(ReplicationMode.FULL)
@@ -252,7 +339,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void metadataOnlySyncEventForTable() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterTableEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterTableEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAlterTableEvent);
 
@@ -261,7 +348,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAlterTableEvent.getEventType()).thenReturn(EventType.ALTER_TABLE);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ALTER_TABLE, TEST_DB, TEST_TABLE)
+        .builder(EventType.ALTER_TABLE, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .environmentContext(EMPTY_MAP)
         .parameters(PARAMETERS)
         .replicationMode(ReplicationMode.METADATA_UPDATE)
@@ -275,7 +362,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void metadataOnlySyncEventForTableWithNullLocation() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterTableEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryAlterTableEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryAlterTableEvent);
 
@@ -283,7 +370,7 @@ public class MessageReaderAdapterTest {
     when(apiaryAlterTableEvent.getEventType()).thenReturn(EventType.ALTER_TABLE);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.ALTER_TABLE, TEST_DB, TEST_TABLE)
+        .builder(EventType.ALTER_TABLE, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .environmentContext(EMPTY_MAP)
         .parameters(PARAMETERS)
         .replicationMode(ReplicationMode.FULL)
@@ -297,7 +384,7 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void dropPartitionEvent() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryDropPartitionEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryDropPartitionEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryDropPartitionEvent);
 
@@ -306,7 +393,7 @@ public class MessageReaderAdapterTest {
     when(apiaryDropPartitionEvent.getEventType()).thenReturn(EventType.DROP_PARTITION);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.DROP_PARTITION, TEST_DB, TEST_TABLE)
+        .builder(EventType.DROP_PARTITION, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .partitionColumns(new ArrayList<String>(PARTITION_KEYS_MAP.keySet()))
         .partitionValues(PARTITION_VALUES)
         .deleteData(true)
@@ -327,7 +414,7 @@ public class MessageReaderAdapterTest {
         .collect(LinkedHashMap::new,
             (m, i) -> m.put(partitionKeys.get(i).getName(), partitionValues.get(0).getValues().get(i)), Map::putAll);
 
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryInsertTableEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryInsertTableEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryInsertTableEvent);
 
@@ -335,7 +422,7 @@ public class MessageReaderAdapterTest {
     when(apiaryInsertTableEvent.getEventType()).thenReturn(EventType.INSERT);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.INSERT, TEST_DB, TEST_TABLE)
+        .builder(EventType.INSERT, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .partitionColumns(new ArrayList<String>(PARTITION_KEYS_MAP.keySet()))
         .partitionValues(PARTITION_VALUES)
         .environmentContext(EMPTY_MAP)
@@ -350,13 +437,13 @@ public class MessageReaderAdapterTest {
 
   @Test
   public void dropTableEvent() {
-    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryDropTableEvent, 
+    when(messageReader.read()).thenReturn(Optional.of(new MessageEvent(apiaryDropTableEvent,
         Collections.singletonMap(SqsMessageProperty.SQS_MESSAGE_RECEIPT_HANDLE, RECEIPT_HANDLE))));
     configureMockedEvent(apiaryDropTableEvent);
     when(apiaryDropTableEvent.getEventType()).thenReturn(EventType.DROP_TABLE);
 
     MetaStoreEvent expected = MetaStoreEvent
-        .builder(EventType.DROP_TABLE, TEST_DB, TEST_TABLE)
+        .builder(EventType.DROP_TABLE, TEST_DB, TEST_TABLE, REPLICA_DATABASE, REPLICA_TABLE)
         .deleteData(true)
         .environmentContext(EMPTY_MAP)
         .parameters(PARAMETERS)
@@ -383,6 +470,8 @@ public class MessageReaderAdapterTest {
     assertThat(actual.getEventType()).isEqualTo(expected.getEventType());
     assertThat(actual.getDatabaseName()).isEqualTo(expected.getDatabaseName());
     assertThat(actual.getTableName()).isEqualTo(expected.getTableName());
+    assertThat(actual.getReplicaDatabaseName()).isEqualTo(expected.getReplicaDatabaseName());
+    assertThat(actual.getReplicaTableName()).isEqualTo(expected.getReplicaTableName());
     assertThat(actual.getPartitionColumns()).isEqualTo(expected.getPartitionColumns());
     assertThat(actual.getPartitionValues()).isEqualTo(expected.getPartitionValues());
     assertThat(actual.getReplicationMode()).isEqualTo(expected.getReplicationMode());